init: Enable touch buttons and bar light
The touch buttons are controlled by the ADUX1050 Capacitive
Proximity Sensor: this device can get the right calibration
only after the projection unit is up and running, so we need
to bring it up from userspace, after initlight.

Then, the touch buttons bar has got lights under it and these
ones are controlled by the supplementary AS3668 chip: raise
the brightness minimally (it's really enough) so that the
user can visually locate the points where to touch.